separated spec files for blockquote
